Refactor TransientServiceProvider
