+++ tests/DependencyInjection/BitzArt.TransientServiceProvider.Tests/ServiceCollectionExtensionsTests.cs
